Updated documentation file, seperate user commands from developer commands for each game. Initialized files, images, and folder for the game Genshin Impact
