+++ games/genshinimpact/config.py
@@ -1,0 +1,10 @@
+import discord
+
+GENSHIN_IMPACT_CONFIG ={
+    "display_name": "Genshin Impact",
+    "api_url": "https://genshin-impact.fandom.com/api.php",
+    "category": "Characters",
+    "color": 0xFFFFFF, # white color
+    "thumbnail_path": "images/GenshinImpactThumbnail.jpeg",
+    "thumbnail_filename": "GenshinImpactThumbnail.png",
+}
