+++ games/wuwa/__init__.py
@@ -1,5 +1,6 @@
 from .api import get_wuwa_events_async
 from .config import WUWA_CONFIG
-from .commands import register_wuwa_commands, register_wuwa_dev_commands
+from .commands import register_wuwa_commands
+from .dev_commands import register_wuwa_dev_commands
 
 __all__ = ["get_wuwa_events_async", "WUWA_CONFIG", "register_wuwa_commands", "register_wuwa_dev_commands"]
