+++ games/wuwa/config.py
@@ -5,7 +5,6 @@ WUWA_CONFIG = {
     "api_url": "https://wutheringwaves.fandom.com/api.php",
     "category": "Events",
     "color": discord.Color.blue(),
-    "emoji": "🌊",
     "thumbnail_path": "images/WutheringWavesThumbnail.jpeg",
     "thumbnail_filename": "WuWaThumbnail.png",
 }
