+++ games/zzz/__init__.py
@@ -1,5 +1,6 @@
 from .api import get_zzz_events_async
 from .config import ZZZ_CONFIG
-from .commands import register_zzz_commands, register_zzz_dev_commands
+from .commands import register_zzz_commands
+from .dev_commands import register_zzz_dev_commands
 
 __all__ = ["get_zzz_events_async", "ZZZ_CONFIG", "register_zzz_commands", "register_zzz_dev_commands"]
