+++ games/zzz/config.py
@@ -5,7 +5,6 @@ ZZZ_CONFIG = {
     "api_url": "https://zenless-zone-zero.fandom.com/api.php",
     "category": "In-Game_Events",
     "color": discord.Color.orange(),
-    "emoji": "⚡",
     "thumbnail_path": "images/ZenlessZoneZeroThumbnail.png",
     "thumbnail_filename": "ZZZThumbnail.png",
 }
